Fix bug in dialogflow related to add specification intent part 4

+++ chatbot_webservice/utils.py
@@ -22,7 +22,7 @@ class Utils:
             'decade': -1
         }
 
-        return duration[Param.AMOUNT] * mult[duration[Param.UNIT]]
+        return float(duration[Param.AMOUNT] * mult[duration[Param.UNIT]])
 
 
 class LossService:
